+++ src/navigation/AppContainer.ts
@@ -7,10 +7,10 @@ import ProfileStack from "../screens/ProfileStack";
 import { Discover } from "../containers/Discover";
 import { Liked } from "../containers/Liked";
 import { Message } from "../containers/Message";
+import { Chat } from "../containers/Chat";
 import { EditProfileScreen } from "../screens/EditProfileScreen";
 import { EditNameScreen } from "../screens/EditNameScreen";
 import { ProfileDetailScreen } from "../screens/ProfileDetailScreen";
-import { ChatScreen } from "../screens/ChatScreen";
 
 const TabNavigator = createBottomTabNavigator(
   {
@@ -37,7 +37,7 @@ const HomeStack = createStackNavigator({
   EditProfile: EditProfileScreen,
   EditName: EditNameScreen,
   ProfileDetail: ProfileDetailScreen,
-  Chat: ChatScreen
+  Chat
 });
 
 const AuthStack = createStackNavigator({ SignIn: SignInScreen });
