+++ src/screens/ChatScreen.tsx
@@ -1,28 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { GiftedChat } from "react-native-gifted-chat";
 
-export const ChatScreen = () => {
-  const [messages, setMessages] = useState([]);
-
-  const onSend = newMessages => {
-    setMessages(prevMessages => [...newMessages, ...prevMessages]);
-  };
-
-  useEffect(() => {
-    setMessages([
-      {
-        _id: 1,
-        text: "Hello developer",
-        createdAt: new Date(),
-        user: {
-          _id: 2,
-          name: "React Native",
-          avatar: "https://placeimg.com/140/140/any"
-        }
-      }
-    ]);
-  }, []);
-
+export const ChatScreen = ({ messages, onSend }) => {
   return (
     <GiftedChat
       messages={messages}
